Fix cast warning

+++ src/streams.h
@@ -458,7 +458,7 @@ private:
     const int nType;
     const int nVersion;
 
-    FILE* file;	
+    FILE* file;
 
 public:
     CAutoFile(FILE* filenew, int nTypeIn, int nVersionIn) : nType(nTypeIn), nVersion(nVersionIn)
@@ -579,7 +579,7 @@ protected:
     bool Fill() {
         unsigned int pos = nSrcPos % vchBuf.size();
         unsigned int readNow = vchBuf.size() - pos;
-        unsigned int nAvail = vchBuf.size() - (nSrcPos - nReadPos) - nRewind;
+        unsigned int nAvail = (unsigned int)(vchBuf.size() - (nSrcPos - nReadPos) - nRewind);
         if (nAvail < readNow)
             readNow = nAvail;
         if (readNow == 0)
@@ -594,8 +594,14 @@ protected:
     }
 
 public:
-    CBufferedFile(FILE *fileIn, uint64_t nBufSize, uint64_t nRewindIn, int nTypeIn, int nVersionIn) :
-        nType(nTypeIn), nVersion(nVersionIn), nSrcPos(0), nReadPos(0), nReadLimit((uint64_t)(-1)), nRewind(nRewindIn), vchBuf(nBufSize, 0)
+    CBufferedFile(FILE *fileIn, uint64_t nBufSize, uint64_t nRewindIn, int nTypeIn, int nVersionIn)
+    : nType(nTypeIn)
+    , nVersion(nVersionIn)
+    , nSrcPos(0)
+    , nReadPos(0)
+    , nReadLimit((uint64_t)(-1))
+    , nRewind(nRewindIn)
+    , vchBuf((unsigned int)nBufSize, 0) // {0}
     {
         src = fileIn;
     }
